Fix e2db client endpoint sync
The e2db client performed an endpoint sync with the etcd server at
initialization which causes the endpoints to get replaced. This is a
problem when dealing with clusters behind terminating load balancers or
any type of opaque service that will cause the local client endpoints to
be replaced with unreachable addresses. It is for this reason that it is
now disabled by default.