pkg/e2db: fix Delete bug on primary index
Fixed a branch in the Delete logic when the key is a primary index.
Return the rows affected in this case. In the secondary index case,
increment rows affected by 1 for each k/v pair, rather than 1 per key
and 1 per value. Do not return an error if the secondary index key does
not exist, to match the primary index deletion behavior.