e2db: promote nested struct fields when embedded
This represents a change in behavior for e2db's ModelDef - previously
an embedded struct or struct pointer would represent a single field,
this change instead promotes the (non-conflicting) fields of the
embedded struct type into the ModelDef of the parent. If there is an
ambiguity in the nested struct fields that would be promoted (as
determined by reflect.Type.FieldByName), NewModelDef panics.