Pin Go module to etcd 3.4
The etcd Go module requires the pseudo version be used (rather than like
v3.4.0) because the Go module version is not being set to v3 in the
go.mod file. This is not an issue, it just means the pseudo version must
be used instead. This also uses the branch release-3.4 which includes
any backports for the 3.4 release.

Update gossip each iteration while bootstrapping

Fix flag parsing with newer Go versions