Add user-defined services
The embedded etcd package allows for user-defined services to be added. A couple grpc methods have been added to help facilitate CA renewal: cluster health check and server restart.

In order to properly determine the health of a given etcd cluster, the required cluster size must be known. This is so that individual members can reason about health in relation to members (regardless of whether a network partition is occurring). To achieve this, a cluster information key is set in the volatile key namespace (i.e. doesn't survive restores from snapshots) that includes the require cluster size. Since this cannot change, this information can be used by an individual member to determine health since it will always be correct. While the goal is to allow an individual member to respond regardless of network partition, the current implementation conservatively still makes a bogus range query, and therefore will be forwarded to the leader (in the case the member is not the leader).

The ability to restart a server has also been added. This restarts only the etcd server, while leaving other services like the gossip network and snapshotter running (although the snapshotter will not produce new snapshots while restarting). This is exposed via the Restart grpc method to allow for remote restarts of the cluster. This will allow for a service to run on the same machine that supervises renewing CA certificates for a multi-node cluster, by replacing the files and then initiating a restart.