Fix race condition after etcd server start
Just after the etcd server is started, it was possible for the server to
stop and to then call code related to clearing the volatile namepsace.
This would cause a panic and didn't manifest until running the manager
tests (-test.long). While very unlikely to affect normal usage, this was
still a race and impacted the ability to run tests safely, so it was
fixed by protecting unsafe etcd server operations with a mutex.